Changed naming convention for RealTime classes to *CORE_X12DSRealTime*.java, moved PortDescriptor class to a non-RealTime package
