+++ Product/Production/Services/CORE_X12DocumentSubmissionCore/src/main/java/gov/hhs/fha/nhinc/corex12/docsubmission/realtime/adapter/component/proxy/AdapterComponentCORE_X12DSRealTimeProxy.java
@@ -34,7 +34,7 @@ import org.caqh.soap.wsdl.corerule2_2_0.COREEnvelopeRealTimeResponse;
  * @author cmay
  *
  */
-public interface AdapterComponentCORE_X12DocSubmissionProxy {
+public interface AdapterComponentCORE_X12DSRealTimeProxy {
 
     public COREEnvelopeRealTimeResponse realTimeRequest(COREEnvelopeRealTimeRequest msg,
         AssertionType assertion);
