+++ Product/Production/Services/CORE_X12DocumentSubmissionCore/src/main/java/gov/hhs/fha/nhinc/corex12/docsubmission/realtime/adapter/proxy/AdapterCORE_X12DSRealTimeProxyObjectFactory.java
@@ -24,20 +24,25 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package gov.hhs.fha.nhinc.corex12.docsubmission.realtime.outbound;
+package gov.hhs.fha.nhinc.corex12.docsubmission.realtime.adapter.proxy;
 
-import gov.hhs.fha.nhinc.common.nhinccommon.AssertionType;
-import gov.hhs.fha.nhinc.common.nhinccommon.NhinTargetCommunitiesType;
-import gov.hhs.fha.nhinc.common.nhinccommon.UrlInfoType;
-import org.caqh.soap.wsdl.corerule2_2_0.COREEnvelopeRealTimeRequest;
-import org.caqh.soap.wsdl.corerule2_2_0.COREEnvelopeRealTimeResponse;
+import gov.hhs.fha.nhinc.proxy.ComponentProxyObjectFactory;
 
 /**
  * @author cmay
  *
  */
-public interface OutboundCORE_X12DocSubmission {
+public class AdapterCORE_X12DSRealTimeProxyObjectFactory extends ComponentProxyObjectFactory {
 
-    public COREEnvelopeRealTimeResponse realTimeRequest(COREEnvelopeRealTimeRequest msg,
-        AssertionType assertion, NhinTargetCommunitiesType targets, UrlInfoType urlInfo);
+    private static final String CONFIG_FILE_NAME = "CORE_X12DocumentSubmissionProxyConfig.xml";
+    private static final String BEAN_NAME = "adaptercore_x12docsubmission";
+
+    @Override
+    protected String getConfigFileName() {
+        return CONFIG_FILE_NAME;
+    }
+
+    public AdapterCORE_X12DSRealTimeProxy getAdapterCORE_X12DocSubmissionProxy() {
+        return getBean(BEAN_NAME, AdapterCORE_X12DSRealTimeProxy.class);
+    }
 }
