+++ Product/Production/Services/CORE_X12DocumentSubmissionCore/src/main/java/gov/hhs/fha/nhinc/corex12/docsubmission/realtime/inbound/AbstractInboundCORE_X12DSRealTime.java
@@ -24,25 +24,36 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package gov.hhs.fha.nhinc.corex12.docsubmission.realtime.adapter.proxy;
+package gov.hhs.fha.nhinc.corex12.docsubmission.realtime.inbound;
 
-import gov.hhs.fha.nhinc.proxy.ComponentProxyObjectFactory;
+import gov.hhs.fha.nhinc.common.nhinccommon.AssertionType;
+import gov.hhs.fha.nhinc.corex12.docsubmission.realtime.adapter.proxy.AdapterCORE_X12DSRealTimeProxy;
+import gov.hhs.fha.nhinc.corex12.docsubmission.realtime.adapter.proxy.AdapterCORE_X12DSRealTimeProxyObjectFactory;
+import org.caqh.soap.wsdl.corerule2_2_0.COREEnvelopeRealTimeRequest;
+import org.caqh.soap.wsdl.corerule2_2_0.COREEnvelopeRealTimeResponse;
 
 /**
  * @author cmay
  *
  */
-public class AdapterCORE_X12DocSubmissionProxyObjectFactory extends ComponentProxyObjectFactory {
+public abstract class AbstractInboundCORE_X12DSRealTime implements InboundCORE_X12DSRealTime {
 
-    private static final String CONFIG_FILE_NAME = "CORE_X12DocumentSubmissionProxyConfig.xml";
-    private static final String BEAN_NAME = "adaptercore_x12docsubmission";
+    private AdapterCORE_X12DSRealTimeProxyObjectFactory adapterFactory;
+
+    abstract COREEnvelopeRealTimeResponse processCORE_X12DocSubmission(COREEnvelopeRealTimeRequest body,
+        AssertionType assertion);
 
     @Override
-    protected String getConfigFileName() {
-        return CONFIG_FILE_NAME;
+    public COREEnvelopeRealTimeResponse realTimeRequest(COREEnvelopeRealTimeRequest body,
+        AssertionType assertion) {
+
+        return processCORE_X12DocSubmission(body, assertion);
     }
 
-    public AdapterCORE_X12DocSubmissionProxy getAdapterCORE_X12DocSubmissionProxy() {
-        return getBean(BEAN_NAME, AdapterCORE_X12DocSubmissionProxy.class);
+    protected COREEnvelopeRealTimeResponse sendToAdapter(COREEnvelopeRealTimeRequest request,
+        AssertionType assertion) {
+
+        AdapterCORE_X12DSRealTimeProxy proxy = adapterFactory.getAdapterCORE_X12DocSubmissionProxy();
+        return proxy.realTimeRequest(request, assertion);
     }
 }
